impl(server): Phase 2A r2 commit 3 — DI migration to Tag-yield handlers (#542)
Migrate every RPC handler from `createXHandlers({deps})` factories to
top-level `xHandlers: RpcMethodRegistry` consts whose binding bodies
pull service Tags via `yield* XServiceTag`. The dispatch runtime carries
`FullLive` so handler bodies resolve R-channel requirements at request
time without per-frame `Effect.provide`.

Per the Phase 2A r2 v5 plan §3 handler audit (P0 consequence):
`network/handlers/auth.handlers.ts` splits into:
- `task/handlers/connect.handlers.ts` (the Connect verb; task-tier since
  it yields ConversationServiceTag + SessionValidatorTag).
- `identity/handlers/agents-lookup.handlers.ts` (AgentsLookup,
  AgentsLookupByName, AgentsList; identity-tier — pure-read agent
  lookups that yield only DbTag + AuthServiceTag).

`app/server.ts` boot site:
- `Effect.runSync(resolveServices.pipe(Effect.provide(FullLive)))`
  duplicate-construction removed; the single `dispatchRuntime` carries
  `FullLive` and `resolveServices` is run inside it.
- `appHost.setConversationService(...)` post-construction wiring moves
  into `Layer.effectDiscard` (Q1 resolution: option (a) chosen).
- Eight factory calls collapse to eight registry imports.

`task/handlers/presence.handlers.test.ts` shifts to Layer-based fakes
(`Layer.mergeAll` of `Layer.succeed(DbTag, db)` + the live
`PresenceService`).

Barrel `index.ts` renames `create*Handlers` exports to `*Handlers`
(no external consumers per audit; verified `grep` returns zero hits
outside `packages/server/src/`).
