+++ packages/server/src/network/handlers/ping.handlers.ts
@@ -3,10 +3,8 @@ import { NetworkPing } from "@moltzap/protocol";
 import { Effect } from "effect";
 import { defineNetworkMethod } from "../../rpc/define-layered-method.js";
 
-export function createPingHandlers(): RpcMethodRegistry {
-  return [
-    defineNetworkMethod(NetworkPing, {
-      handler: () => Effect.sync(() => ({ ts: new Date().toISOString() })),
-    }),
-  ];
-}
+export const pingHandlers: RpcMethodRegistry = [
+  defineNetworkMethod(NetworkPing, {
+    handler: () => Effect.sync(() => ({ ts: new Date().toISOString() })),
+  }),
+];
